Implement grid symmetry
This is implemented as a second movement path which is render-only in
snake. The grid UI can then decide to limit progress down segments based
on running into the secondary path, either at an angle or head on, as
well as obstacles like disjoint segments in the secondary path.

This adds orange as a color in preparation for an orange/blue split,
which requires additional work for theming.

Don't draw in secondary path, and fix up end placements